set unit test types, remove obsolete code

+++ WebApi.NetCore/WebApi.NetCore.Api/Database/MongoDbProvider.cs
@@ -1,10 +1,11 @@
 ﻿namespace WebApi.NetCore.Api.Database
 {
     using System.Linq.Expressions;
+    using Generic.Base.Api.Database;
+    using Generic.Base.Api.Exceptions;
+    using Generic.Base.Api.Transformer;
     using MongoDB.Driver;
     using WebApi.NetCore.Api.Contracts.Configuration;
-    using WebApi.NetCore.Api.Contracts.Database;
-    using WebApi.NetCore.Api.Exceptions;
 
     /// <summary>
     ///     An generic class for mongo db based providers.
@@ -12,23 +13,20 @@
     /// <typeparam name="TEntry">The type of the entry.</typeparam>
     /// <typeparam name="TDatabaseEntry">The type of the database entry.</typeparam>
     /// <typeparam name="TDatabaseConfiguration">The type of the database configuration.</typeparam>
-    public abstract class
+    public class
         MongoDbProvider<TEntry, TDatabaseEntry, TDatabaseConfiguration> : IProvider<TEntry, IClientSessionHandle>
-        where TDatabaseConfiguration : IDatabaseConfiguration where TDatabaseEntry : IDatabaseEntry
+        where TDatabaseConfiguration : IDatabaseConfiguration where TDatabaseEntry : IIdEntry
     {
         private readonly IMongoCollection<TDatabaseEntry> collection;
-        private readonly Func<TEntry, TDatabaseEntry> toDatabaseEntry;
-        private readonly Func<TDatabaseEntry, TEntry> toEntry;
+        private readonly IProviderEntryTransformer<TEntry, TDatabaseEntry> transformer;
 
-        protected MongoDbProvider(
+        public MongoDbProvider(
             IMongoClient mongoClient,
             TDatabaseConfiguration databaseConfiguration,
-            Func<TEntry, TDatabaseEntry> toDatabaseEntry,
-            Func<TDatabaseEntry, TEntry> toEntry
+            IProviderEntryTransformer<TEntry, TDatabaseEntry> transformer
         )
         {
-            this.toDatabaseEntry = toDatabaseEntry;
-            this.toEntry = toEntry;
+            this.transformer = transformer;
             this.collection = mongoClient.GetDatabase(databaseConfiguration.DatabaseName)
                 .GetCollection<TDatabaseEntry>(databaseConfiguration.CollectionName);
         }
@@ -39,20 +37,21 @@
         /// <param name="entry">The entry.</param>
         /// <param name="cancellationToken">Indicates that the start process has been aborted.</param>
         /// <param name="transactionHandle">Handle for database transactions.</param>
-        /// <returns>A <see cref="Task" /> whose result indicates success.</returns>
-        public async Task CreateAsync(
+        /// <returns>A <see cref="Task" /> whose result is the created entry.</returns>
+        public async Task<TEntry> CreateAsync(
             TEntry entry,
             CancellationToken cancellationToken,
             ITransactionHandle<IClientSessionHandle> transactionHandle
         )
         {
-            var databaseEntry = this.toDatabaseEntry(entry);
+            var databaseEntry = this.transformer.Transform(entry);
             try
             {
                 await this.collection.InsertOneAsync(
                     transactionHandle.ClientSessionHandle,
                     databaseEntry,
                     cancellationToken: cancellationToken);
+                return this.transformer.Transform(databaseEntry);
             }
             catch (MongoWriteException e) when (e.WriteError.Code == 11000)
             {
@@ -71,21 +70,26 @@
         /// <summary>
         ///     Delete an entry by its id used in the application.
         /// </summary>
-        /// <param name="applicationId">The id of the entry used in the application.</param>
+        /// <param name="id">The id of the entry used in the application.</param>
         /// <param name="cancellationToken">The cancellation token.</param>
         /// <param name="transactionHandle">The transaction handle.</param>
         /// <returns>A <see cref="Task" /> whose result indicates success.</returns>
         public async Task DeleteAsync(
-            string applicationId,
+            string id,
             CancellationToken cancellationToken,
             ITransactionHandle<IClientSessionHandle> transactionHandle
         )
         {
+            if (string.IsNullOrWhiteSpace(id))
+            {
+                throw new ArgumentNullException(nameof(id));
+            }
+
             var result = await this.collection.DeleteOneAsync(
                 transactionHandle.ClientSessionHandle,
                 doc => string.Equals(
-                    doc.ApplicationId,
-                    applicationId,
+                    doc.Id,
+                    id,
                     StringComparison.OrdinalIgnoreCase),
                 cancellationToken: cancellationToken);
             if (!result.IsAcknowledged || result.DeletedCount == 0)
@@ -110,27 +114,32 @@
                 doc => true,
                 cancellationToken: cancellationToken);
             var result = await cursor.ToListAsync(cancellationToken);
-            return result.Select(this.toEntry).ToArray();
+            return result.Select(this.transformer.Transform).ToArray();
         }
 
         /// <summary>
         ///     Read an entry by its id used in the application.
         /// </summary>
-        /// <param name="applicationId">The id of the entry used in the application.</param>
+        /// <param name="id">The id of the entry used in the application.</param>
         /// <param name="cancellationToken">The cancellation token.</param>
         /// <param name="transactionHandle">The transaction handle.</param>
         /// <returns>A <see cref="Task{T}" /> whose result is the found entry.</returns>
         public async Task<TEntry> ReadByIdAsync(
-            string applicationId,
+            string id,
             CancellationToken cancellationToken,
             ITransactionHandle<IClientSessionHandle> transactionHandle
         )
         {
+            if (string.IsNullOrWhiteSpace(id))
+            {
+                throw new ArgumentNullException(nameof(id));
+            }
+
             var result = await this.collection.FindAsync(
                 transactionHandle.ClientSessionHandle,
                 doc => string.Equals(
-                    doc.ApplicationId,
-                    applicationId,
+                    doc.Id,
+                    id,
                     StringComparison.OrdinalIgnoreCase),
                 cancellationToken: cancellationToken);
             var entry = result.FirstOrDefault(cancellationToken);
@@ -139,7 +148,7 @@
                 throw new NotFoundException($"Cannot read {nameof(TDatabaseEntry)} entry due to its unknown id.");
             }
 
-            return this.toEntry(entry);
+            return this.transformer.Transform(entry);
         }
 
         /// <summary>
@@ -149,11 +158,41 @@
         /// <param name="cancellationToken">The cancellation token.</param>
         /// <param name="transactionHandle">The transaction handle.</param>
         /// <returns>A <see cref="Task" /> whose result indicates success.</returns>
-        public abstract Task UpdateAsync(
+        public async Task UpdateAsync(
             TEntry entry,
             CancellationToken cancellationToken,
             ITransactionHandle<IClientSessionHandle> transactionHandle
-        );
+        )
+        {
+            var databaseEntry = this.transformer.Transform(entry);
+            try
+            {
+                var result = await this.collection.ReplaceOneAsync(
+                    transactionHandle.ClientSessionHandle,
+                    Builders<TDatabaseEntry>.Filter.Eq(
+                        doc => doc.Id,
+                        databaseEntry.Id),
+                    databaseEntry,
+                    new ReplaceOptions
+                    {
+                        Collation = new Collation(
+                            "en",
+                            strength: new Optional<CollationStrength?>(CollationStrength.Secondary))
+                    },
+                    cancellationToken);
+                if (!result.IsAcknowledged || result.MatchedCount == 0)
+                {
+                    throw new NotFoundException(
+                        $"Cannot replace {nameof(TDatabaseEntry)} entry: Filter does not match.");
+                }
+            }
+            catch (MongoWriteException e) when (e.WriteError.Code == 121)
+            {
+                throw new BadRequestException(
+                    $"Cannot update {nameof(TDatabaseEntry)} entry due to validation failure.",
+                    e);
+            }
+        }
 
         /// <summary>
         ///     Updates the specified entry.
